+++ sphinx/api/application.rst
@@ -1,0 +1,7 @@
+.. _application:
+
+
+Application
+===========
+
+..  automodule:: datacatalog.application
