+++ sphinx/api/handlers.rst
@@ -4,8 +4,8 @@
 Handlers
 ========
 
-.. automodule:: datacatalog.action_api
+.. automodule:: datacatalog.handlers.action_api
 
-.. automodule:: datacatalog.index
+.. automodule:: datacatalog.handlers.index
 
-.. automodule:: datacatalog.search
+.. automodule:: datacatalog.handlers.systemhealth
