--- sphinx/api/plugin_specs.rst
@@ -1,12 +1,0 @@
-.. _plugin_specs:
-
-
-Plugin Specifications
-=====================
-
-.. automodule:: datacatalog.plugin_specs
-.. automodule:: datacatalog.plugin_specs._facet_search
-.. automodule:: datacatalog.plugin_specs._full_text_search
-.. automodule:: datacatalog.plugin_specs._markers
-.. automodule:: datacatalog.plugin_specs._schema
-.. automodule:: datacatalog.plugin_specs._storage
