+++ sphinx/index.rst
@@ -4,6 +4,7 @@
     :hidden:
 
     api_docs
+    plugins
     schemas
     sphinx_101
     todo_list
