+++ src/datacatalog/default_plugins/file_storage_config_schema.yml
