+++ src/datacatalog/default_plugins/in_memory_search_config_schema.yml
